backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.12
Flask==3.1.1
flask-cors==6.0.0
Flask-PyMongo==3.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.29.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.0
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
Werkzeug==3.1.3
